Allow user to provide MD5 hash for Gravatar
Calculating the Gravatar URL based on the email address
in Javascript is convenient, but having the email address
in plaintext in the resulting HTML file might not be what
everybody wants.

This adds support for providing the MD5 hash value via
the email_md5 entry in _config.yml.

As a side-effect, we don't need to include md5.js if the
hash value is already known.